chore: update permission

+++ src/features/notes/server/actions/remove-note.ts
@@ -1,19 +1,24 @@
 "use server";
 
-import {
-  requireAuth,
-  requireNoteEdit,
-  requireWorkspaceMember,
-} from "@/lib/permissions";
+import { requireAuth } from "@/lib/permissions";
 import { deleteNote } from "../mutations/delete-note";
+import { requirePermission } from "@/lib/authz";
+import { fail, ok, Result } from "@/lib/types";
 
-export async function removeNote(noteId: string) {
-  await requireNoteEdit(noteId);
+export async function removeNote(
+  noteId: string,
+  workspaceId: string,
+): Promise<Result<string>> {
+  // Auth
+  await requireAuth();
 
+  // Authz
+  await requirePermission(workspaceId, "note:delete");
+
+  // DB Action
   const success = await deleteNote(noteId);
-  if (!success) {
-    return { success, message: "Fail to delete note" };
-  }
 
-  return { success, message: "Successfuly note deleted" };
+  // Response
+  if (!success) return fail("Fail to delet note");
+  return ok("Successfully note deleted");
 }
